adjust entity relationship

+++ cqjz/src/main/java/com/manage/news/jpa/kernel/entity/Menu.java
@@ -22,10 +22,10 @@ public class Menu {
 
     @Id
     @GeneratedValue(generator = "seq_menu", strategy = GenerationType.SEQUENCE)
-    private Integer id;
+    private Long id;
 
     @Column(name = "parent_id", insertable = false, updatable = false)
-    private Integer parentId;
+    private Long parentId;
 
     @Column(name = "sequence", length = 3)
     private Integer sequence = 0;
@@ -50,19 +50,19 @@ public class Menu {
     @OrderBy("sequence asc")
     private List<Menu> childrens;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getParentId() {
+    public Long getParentId() {
         return parentId;
     }
 
-    public void setParentId(Integer parentId) {
+    public void setParentId(Long parentId) {
         this.parentId = parentId;
     }
